Change config_key_vertical_gap
I can't live with this much gap between my vertical rows. I assume this will fuck up the 5 row layout settings below which was meant to keep the 4 row and 5 row keyboard at the same height. But who would want that if it messes up the aesthetics of the layout.